requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
tensorflow
Pillow